feat: add feature save meal

+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/data/nutrition/datasource/local/MealEntity.kt
@@ -9,7 +9,7 @@ data class MealEntity(
     val id: String,
     val userId: String,
     val dishLabel: String,
-    val servingSize: Float,
+    val servingSize: Double,
     val components: Map<String, NutritionItem>,
     val totalNutrition: TotalNutrition,
     val date: String,
