+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/data/nutrition/datasource/remote/dto/MealDto.kt
@@ -7,7 +7,7 @@ data class MealDto(
     val id: String,
     val userId: String,
     val dishLabel: String,
-    val servingSize: Float,
+    val servingSize: Double,
     val components: Map<String, NutritionItem>,
     val totalNutrition: TotalNutrition,
     val date: String,
